ACON Activation batch-size 1 bug patch (#2901)
* ACON Activation batch-size 1 bug path

This is not a great solution to https://github.com/nmaac/acon/issues/4 but it's all I could think of at the moment.

WARNING: YOLOv5 models with MetaAconC() activations are incapable of running inference at batch-size 1 properly due to a known bug in https://github.com/nmaac/acon/issues/4 with no known solution.

* Update activations.py

* Update activations.py

* Update activations.py

* Update activations.py