Allow list for Comet artifact class 'names' field (#9654)
* Update __init__.py

In the Comet logger, when I run train.py, it wants to download the data artifact. It was requiring me to format the 'names' field in the data artifact metadata as a dictionary, so I've changed this so that it also accepts a list.

Signed-off-by: KristenKehrer <34010022+KristenKehrer@users.noreply.github.com>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update utils/loggers/comet/__init__.py

Co-authored-by: Dhruv Nair <dhruv.nair@gmail.com>
Signed-off-by: KristenKehrer <34010022+KristenKehrer@users.noreply.github.com>

Signed-off-by: KristenKehrer <34010022+KristenKehrer@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dhruv Nair <dhruv.nair@gmail.com>
Co-authored-by: Glenn Jocher <glenn.jocher@ultralytics.com>